backend/requirements.txt
fastapi
uvicorn
PyMuPDF
sentence-transformers
langchain
google-generativeai
numpy
python-multipart
langchain-community
qdrant-client
python-dotenv
psycopg2-binary 
passlib 
python-jose[cryptography]
httpx
sqlalchemy
authlib
itsdangerous